Issue #258: Move all regression tests, apart from mixed lang, out of /examples into /test
