--- examples/sim_exit/tests/mem.out
@@ -1,40 +1,0 @@
-
-
-Running test 1/8:test_yield_list
-
-End of test
-
-
-Running test 2/8:test_function_not_a_coroutine
-
-End of test
-
-
-Running test 3/8:test_coroutine_kill
-
-End of test
-
-
-Running test 4/8:test_adding_a_coroutine_without_starting
-
-End of test
-
-
-Running test 5/8:test_function_not_a_coroutine_fork
-
-End of test
-
-
-Running test 6/8:test_function_reentrant_clock
-
-End of test
-
-
-Running test 7/8:simple_test
-
-End of test
-
-
-Running test 8/8:discover_module_values
-
-End of test
