+++ tests/Makefile
@@ -1,6 +1,5 @@
 ###############################################################################
-# Copyright (c) 2013 Potential Ventures Ltd
-# Copyright (c) 2013 SolarFlare Communications Inc
+# Copyright (c) 2015 Potential Ventures Ltd
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -10,8 +9,7 @@
 # * Redistributions in binary form must reproduce the above copyright
 # notice, this list of conditions and the following disclaimer in the
 # documentation and/or other materials provided with the distribution.
-# * Neither the name of Potential Ventures Ltd,
-#       SolarFlare Communications Inc nor the
+# * Neither the name of Potential Ventures Ltd nor the
 # names of its contributors may be used to endorse or promote products
 # derived from this software without specific prior written permission.
 #
@@ -30,15 +28,14 @@
 COCOTB?=$(shell pwd)/../
 export COCOTB
 
-EXAMPLES := adder/tests \
-	    endian_swapper/tests \
+REGRESSIONS :=  $(shell ls test_cases/)
 
-.PHONY: $(EXAMPLES)
+.PHONY: $(REGRESSIONS)
 
-all: $(EXAMPLES)
+all: $(REGRESSIONS)
 
-$(EXAMPLES):
-	@cd $@ && $(MAKE)
+$(REGRESSIONS):
+	cd test_cases/$@ && $(MAKE)
 
 clean:
-	$(foreach TEST, $(EXAMPLES), $(MAKE) -C $(TEST) clean;)
+	$(foreach TEST, $(REGRESSIONS), $(MAKE) -C test_cases/$(TEST) clean;)
