+++ tests/designs/close_module/close_module.v
