+++ tests/designs/sample_module/sample_module.vhdl
